pyproject.toml
[project]
name = "local-pdf-translator"
version = "0.1.0"
description = "Local browser-based PDF translation app with DeepSeek OCR and MLX translation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.7.0",
  "pypdf>=4.2.0",
  "pdfplumber>=0.11.0",
  "pypdfium2>=4.30.0",
  "pillow>=10.3.0",
  "langdetect>=1.0.9",
  "markdown>=3.6",
  "weasyprint>=62.3",
]

[project.optional-dependencies]
mlx = [
  "mlx-lm>=0.31.0",
]
deepseek_ocr = [
  "mlx-vlm>=0.4.4",
]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "backend"}

[tool.setuptools.packages.find]
where = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
strict = false
